Fix error when hex-regex does not match
This occurs with shorthand-hex values like '#ddd', but in general I think
its better to handle the case of incorrect data being passed to the
brightness function in a more robust way. This is achieved by returning
null if the regex matching does not work and handling that null value
explicitly on the calling side as well.